+++ src/Views/src/Loader/PathParser.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views\Loader;
@@ -19,16 +12,10 @@ use Spiral\Views\LoaderInterface;
  */
 final class PathParser
 {
-    /** @var string */
-    private $defaultNamespace;
-
-    /** @var string */
-    private $extension;
-
-    public function __construct(string $defaultNamespace, string $extension)
-    {
-        $this->defaultNamespace = $defaultNamespace;
-        $this->extension = $extension;
+    public function __construct(
+        private readonly string $defaultNamespace,
+        private readonly string $extension
+    ) {
     }
 
     public function getExtension(): string
@@ -41,13 +28,13 @@ final class PathParser
      */
     public function match(string $filename): bool
     {
-        $extension = substr($filename, -strlen($this->extension) - 1);
-        return strtolower($extension) === ".{$this->extension}";
+        $extension = \substr($filename, -\strlen($this->extension) - 1);
+
+        return \strtolower($extension) === \sprintf('.%s', $this->extension);
     }
 
     /**
      * Parse view path and extract name, namespace and basename information.
-     *
      *
      * @throws PathException
      */
@@ -56,33 +43,33 @@ final class PathParser
         $this->validatePath($path);
 
         //Cutting extra symbols (see Twig)
-        $filename = preg_replace(
+        $filename = \preg_replace(
             '#/{2,}#',
             '/',
-            str_replace('\\', '/', $path)
+            \str_replace('\\', '/', $path)
         );
 
         $namespace = $this->defaultNamespace;
-        if (strpos($filename, '.') === false) {
+        if (!\str_contains((string) $filename, '.')) {
             //Force default extension
             $filename .= '.' . $this->extension;
         } elseif (!$this->match($filename)) {
             return null;
         }
 
-        if (strpos($filename, LoaderInterface::NS_SEPARATOR) !== false) {
-            [$namespace, $filename] = explode(LoaderInterface::NS_SEPARATOR, $filename);
+        if (\str_contains((string) $filename, LoaderInterface::NS_SEPARATOR)) {
+            [$namespace, $filename] = explode(LoaderInterface::NS_SEPARATOR, (string) $filename);
         }
 
         //Twig like namespaces
-        if (isset($filename[0]) && $filename[0] == '@') {
-            $separator = strpos($filename, '/');
+        if (isset($filename[0]) && $filename[0] === '@') {
+            $separator = \strpos($filename, '/');
             if ($separator === false) {
-                throw new PathException(sprintf('Malformed view path"%s" (expecting "@namespace/name").', $path));
+                throw new PathException(\sprintf('Malformed view path "%s" (expecting "@namespace/name").', $path));
             }
 
-            $namespace = substr($filename, 1, $separator - 1);
-            $filename = substr($filename, $separator + 1);
+            $namespace = \substr($filename, 1, $separator - 1);
+            $filename = \substr($filename, $separator + 1);
         }
 
         return new ViewPath(
@@ -111,12 +98,12 @@ final class PathParser
             throw new PathException('A view path is empty');
         }
 
-        if (false !== strpos($path, "\0")) {
+        if (\str_contains($path, "\0")) {
             throw new PathException('A view path cannot contain NUL bytes');
         }
 
-        $path = ltrim($path, '/');
-        $parts = explode('/', $path);
+        $path = \ltrim($path, '/');
+        $parts = \explode('/', $path);
         $level = 0;
 
         foreach ($parts as $part) {
@@ -127,7 +114,7 @@ final class PathParser
             }
 
             if ($level < 0) {
-                throw new PathException(sprintf(
+                throw new PathException(\sprintf(
                     'Looks like you try to load a view outside configured directories (%s)',
                     $path
                 ));