+++ src/Views/src/Traits/ProcessorTrait.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views\Traits;
@@ -18,7 +11,7 @@ use Spiral\Views\ViewSource;
 trait ProcessorTrait
 {
     /** @var ProcessorInterface[] */
-    private $processors = [];
+    private array $processors = [];
 
     /**
      * Process given view source using set of associated processors.