+++ src/Views/src/ViewCache.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views;
@@ -15,12 +8,8 @@ use Spiral\Views\Exception\CacheException;
 
 final class ViewCache
 {
-    /** @var array */
-    private $cache = [];
+    private array $cache = [];
 
-    /**
-     * @param ContextInterface|null $context
-     */
     public function reset(ContextInterface $context = null): void
     {
         if (empty($context)) {
@@ -52,13 +41,12 @@ final class ViewCache
     }
 
     /**
-     *
      * @throws CacheException
      */
     public function get(ContextInterface $context, string $path): ViewInterface
     {
         if (!$this->has($context, $path)) {
-            throw new CacheException("No cache is available for {$path}.");
+            throw new CacheException(\sprintf('No cache is available for %s.', $path));
         }
 
         return $this->cache[$context->getID()][$path];