+++ src/Views/src/ViewsInterface.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Views;
@@ -18,7 +11,6 @@ interface ViewsInterface
     /**
      * Get instance of view class associated with view path (path can include namespace).
      *
-     *
      * @throws ViewException
      */
     public function get(string $path): ViewInterface;
@@ -26,14 +18,12 @@ interface ViewsInterface
     /**
      * Compile one of multiple cache versions for a given view path.
      *
-     *
      * @throws ViewException
      */
     public function compile(string $path);
 
     /**
      * Reset view cache for a given path.
-     *
      *
      * @throws ViewException
      */